pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "flask>=3.1.3",
    "gunicorn>=25.3.0",
    "psycopg2-binary>=2.9.12",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "validators>=0.35.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "pytest-postgresql>=8.0.0",
    "responses>=0.26.0",
    "ruff>=0.15.12",
]

[tool.ruff]
target-version = "py312"
line-length = 100
exclude = ["docs", ".venv"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # bugbear
]
